REST API: add `publicize_enabled` flag in site options
Summary:
This patch seeks to add a `publicize_permanently_disabled` flag in the `options` object returned as part of the site object. This is part of the `me/sites` and `site/$site` return in the REST API. There are 2 statuses:

* `true` means that the site is never allowed to publicize (because it's an a8c private site, because it's banned or because it's an invalid blog ID)
* `false` means the site is not disallowed

The end goal is to able to provide users with information about whether or not publicize is enabled on their sites without needing to make a request to fetch publicize connections.

This also involved a small refactor of the `disabled` code in Publicize Next.

Merges r124967-wpcom.